Precio con 2 decimales y signo S/ (precioHora int->double)
- Cancha.precioHora pasa de int a double; getters, copyWith, fromJson/toJson
  y precioReferencial ajustados.
- Repos (Supabase) y Club.precioDesde manejan double.
- Formularios de registrar/editar: teclado decimal, parseo con coma->punto,
  inicializacion con 2 decimales.
- Todas las vistas muestran el precio con toStringAsFixed(2) y signo S/.
- Reservas usan precioHora.round() para sena/precio internos (siguen en int).

+++ lib/data/canchas_repo.dart
@@ -96,7 +96,7 @@ class CanchasRepo {
         club: (r['club'] ?? '') as String,
         distrito: _enumDistrito(r['distrito'] as String?),
         deporte: _enumDeporte(r['deporte'] as String?),
-        precioHora: ((r['precio_hora'] ?? 100) as num).toInt(),
+        precioHora: ((r['precio_hora'] ?? 100) as num).toDouble(),
         ubicacion: LatLng(
           ((r['lat'] ?? -12.108) as num).toDouble(),
           ((r['lng'] ?? -76.978) as num).toDouble(),
